REFACTOR: chat/translation record DTO를 dto 하위 패키지로 분리
TranslationResult, DeepL/Gemini 요청·응답 record를 chat/translation/dto로 이동해
provider 구현체와 DTO를 구분. 다른 feature 패키지의 dto 서브패키지 컨벤션에 맞춤.
