+++ src/main/java/com/deoham/chat/translation/DeepLTranslationProvider.java
@@ -1,5 +1,8 @@
 package com.deoham.chat.translation;
 
+import com.deoham.chat.translation.dto.DeepLTranslateRequest;
+import com.deoham.chat.translation.dto.DeepLTranslateResponse;
+import com.deoham.chat.translation.dto.TranslationResult;
 import com.deoham.global.config.DeepLProperties;
 import com.deoham.global.exception.BusinessException;
 import com.deoham.global.exception.ErrorCode;
