+++ src/main/java/com/deoham/chat/translation/TranslationProvider.java
@@ -1,5 +1,7 @@
 package com.deoham.chat.translation;
 
+import com.deoham.chat.translation.dto.TranslationResult;
+
 public interface TranslationProvider {
 
     TranslationResult translate(String text, String targetLanguage);
