+++ src/main/java/com/deoham/chat/translation/dto/GeminiGenerateContentRequest.java
@@ -1,4 +1,4 @@
-package com.deoham.chat.translation;
+package com.deoham.chat.translation.dto;
 
 import java.util.List;
