+++ src/main/java/com/deoham/chat/translation/dto/TranslationResult.java
@@ -1,4 +1,4 @@
-package com.deoham.chat.translation;
+package com.deoham.chat.translation.dto;
 
 public record TranslationResult(String translatedText, String modelVersion) {
 }
