+++ src/test/java/com/deoham/chat/translation/GeminiTranslationProviderTest.java
@@ -1,5 +1,6 @@
 package com.deoham.chat.translation;
 
+import com.deoham.chat.translation.dto.TranslationResult;
 import com.deoham.global.config.GeminiProperties;
 import com.deoham.global.exception.BusinessException;
 import com.deoham.global.exception.ErrorCode;
